requirements.txt
pandas
openpyxl
python-pptx
requests
matplotlib
